+++ Backend/app/src/main/java/Backend/Railroad.java
@@ -1,21 +1,40 @@
 package Backend;
-import lombok.Getter;
-import lombok.Setter;
 
 public class Railroad extends Property{
 
 
-    @Getter @Setter private boolean hasTrainStation;
-
     /**
-     * Constructor
      *
-     * @param name     Name of the space
-     * @param location The position on the gameboard
+     * @param name Name of the Railroad
+     * @param location Position on the Board
+     * @param value The price to purchase the Property
+     * @param rent Cost for landing on the Property
      */
-    public Railroad(String name, int location, boolean tStation) {
-        super(name, location);
-        this.hasTrainStation = tStation;
+    public Railroad(String name, int location, int value, int rent) {
+        super(name, location, value, rent);
     }
 
+    /**
+     *
+     * @param property  Checks the rent value
+     * @return Value of rent
+     */
+    @Override
+    public int getRent(Property property) {
+        if(Owner.isPlayer(getOwner())){
+            if (((Player)getOwner()).railRoadsOwned((Railroad) property) == 1){
+                return 25;
+            }
+            if (((Player)getOwner()).railRoadsOwned((Railroad) property) == 2){
+                return 50;
+            }
+            else if (((Player)getOwner()).railRoadsOwned((Railroad) property) == 3){
+                return 100;
+            }
+            if (((Player)getOwner()).railRoadsOwned((Railroad) property) == 4){
+                return 200;
+            }
+        }
+        return super.getRent(property);
+    }
 }